Merged PR 172812: Added attribute parsing for items and ads
### Summary of Changes
Attribute set parsed solely from `product_catalog` table that will be common between items (and therefore customer histories) and ads for ad relevance modelling. New `task_parse_attributes.py` script not yet scheduled as part of production job - the primary purpose of this PR is to get the new `TargetingAttributes` column into the control sheet and control sheet tables.

### Reason for Changes
Item and ad attributes/labels are currently inferred and based on a mixture of fields, which are maintained to varying standards. Focussing on attributes that are single-source (product_catalog) and reflect the filters that are used on site, is intended as a simpler, clearner approach, that will lower the risk of mistargeting.

### Testing Conducted
- `task_load_control_sheet.py` tested interactively using a 'dev' version of the control sheet with the new column added - no issues.
- `next-ads` dev/preprod job end to end run: see [here](https://adb-6188831950334199.19.azuredatabricks.net/jobs/518755454712672/runs/73492456655716?o=6188831950334199)
- pytest integration tests all passed prior to deployment:
![image.png](https://dev.azure.com/Next-Technology/99c4d67f-45b0-465f-8499-ea3f5852ede1/_apis/git/repositories/73b8c6cc-cd84-4245-b743-7e437ae2b2ed/pullRequests/172812/attachments/image.png)

### Associated Risks and Assumptions
- In addition to the testing conduicted to ensure that there are no operational issues introducted by the new column, the additional column is not yet referenced in any operational part of the codebase, therefore should not be able to cause any regressions, and the additional keys in the config are additional and separate from any other sections of the config.