Merged PR 179023: Add check for tables existence before assert_pk()
# Purpose of this PR

The asser_pk() for loop assumes all write tables have a `warehouse` table equivalent which is not always the case e.g. when we are slowly rolling out a new feature we often create the dev tables first but not the warehouse equivalent tables. This causes an `AnalysisException` when asserting PK's for a table which doesn't exist.

![image (2).png](https://dev.azure.com/Next-Technology/99c4d67f-45b0-465f-8499-ea3f5852ede1/_apis/git/repositories/73b8c6cc-cd84-4245-b743-7e437ae2b2ed/pullRequests/179023/attachments/image%20%282%29.png)

# File changes in this PR
* 'task_qa.py` - Add `spark.catalog.tableExists(tbl_mapped)`  check before conducting the assert_pk() function, if table does not exist print message informing it is being skipped.

# Testing changes in this PR
Local testing has been conducted in both dev/prod environments. This change/script does not have any downstream dependencies so extensive testing is unnesseary.